+++ public/controllers/management/components/management/common/file-editor.tsx
@@ -12,10 +12,6 @@
 import React, { Component, Fragment } from 'react';
 
 import { connect } from 'react-redux';
-import {
-  cleanInfo,
-  updateFileContent
-} from '../../../../../redux/actions/rulesetActions';
 
 import 'brace/theme/textmate';
 // Eui components
@@ -34,8 +30,8 @@ import {
   EuiPanel,
 } from '@elastic/eui';
 
-import { resourceDictionary, RulesetHandler, RulesetResources } from './utils/ruleset-handler';
-import validateConfigAfterSent from './utils/valid-configuration';
+import { resourceDictionary, ResourcesHandler } from './resources-handler';
+import validateConfigAfterSent from './valid-configuration';
 
 import { getToasts } from '../../../../../kibana-services';
 import { updateWazuhNotReadyYet } from '../../../../../redux/actions/appStateActions';
@@ -47,14 +43,13 @@ import 'brace/mode/xml';
 import 'brace/snippets/xml';
 import 'brace/ext/language_tools';
 import "brace/ext/searchbox";
-import { showFlyoutLogtest } from '../../../../../redux/actions/appStateActions';
 import _ from 'lodash';
 
 import { UI_ERROR_SEVERITIES } from '../../../../../react-services/error-orchestrator/types';
 import { UI_LOGGER_LEVELS } from '../../../../../../common/constants';
 import { getErrorOrchestrator } from '../../../../../react-services/common-services';
 
-class WzRulesetEditor extends Component {
+class WzFileEditor extends Component {
   _isMounted = false;
   constructor(props) {
     super(props);
@@ -68,9 +63,9 @@ class WzRulesetEditor extends Component {
       enableSnippets: true,
       enableLiveAutocompletion: false,
     };
-    this.rulesetHandler = new RulesetHandler(this.props.state.section);
-    const { fileContent, addingRulesetFile } = this.props.state;
-    const { name, content, path } = fileContent ? fileContent : addingRulesetFile;
+    this.resourcesHandler = new ResourcesHandler(this.props.section);
+    const { fileContent, addingFile } = this.props;
+    const { name, content, path } = fileContent ? fileContent : addingFile;
 
     this.state = {
       isSaving: false,
@@ -89,7 +84,7 @@ class WzRulesetEditor extends Component {
   componentWillUnmount() {
     // When the component is going to be unmounted its info is clear
     this._isMounted = false;
-    this.props.cleanInfo();
+    this.props.cleanEditState();
   }
 
   componentDidMount() {
@@ -106,7 +101,7 @@ class WzRulesetEditor extends Component {
     if (!this._isMounted) {
       return;
     }else if(/\s/.test(name)) {
-      this.showToast('warning', 'Warning', `The ${this.props.state.section} name must not contain spaces.`, 3000);
+      this.showToast('warning', 'Warning', `The ${this.props.section} name must not contain spaces.`, 3000);
       return;
     }  
     try {
@@ -114,12 +109,12 @@ class WzRulesetEditor extends Component {
 
       this.setState({ isSaving: true, error: false });
 
-      await this.rulesetHandler.updateFile(name, content, overwrite);
+      await this.resourcesHandler.updateFile(name, content, overwrite);
       try {
         await validateConfigAfterSent();
       } catch (error) {
         const options = {
-          context: `${WzRulesetEditor.name}.save`,
+          context: `${WzFileEditor.name}.save`,
           level: UI_LOGGER_LEVELS.ERROR,
           severity: UI_ERROR_SEVERITIES.BUSINESS,
           error: {
@@ -134,13 +129,13 @@ class WzRulesetEditor extends Component {
 
         let toastMessage;
 
-        if (this.props.state.addingRulesetFile != false) {
+        if (this.props.state.addingFile != false) {
           //remove current invalid file if the file is new.
-          await this.rulesetHandler.deleteFile(name);
+          await this.resourcesHandler.deleteFile(name);
           toastMessage = 'The new file was deleted.';
         } else {
           //restore file to previous version
-          await this.rulesetHandler.updateFile(name, this.state.initContent, overwrite);
+          await this.resourcesHandler.updateFile(name, this.state.initContent, overwrite);
           toastMessage = 'The content file was restored to previous state.';
         }
 
@@ -156,14 +151,20 @@ class WzRulesetEditor extends Component {
       });
 
     } catch (error) {
+      let errorMessage;
+      if (error instanceof Error) {
+        errorMessage = error.details
+          ? error.details
+          : String(error);
+      }
       this.setState({ error, isSaving: false });
       const options = {
-        context: `${WzRulesetEditor.name}.save`,
+        context: `${WzFileEditor.name}.save`,
         level: UI_LOGGER_LEVELS.ERROR,
         severity: UI_ERROR_SEVERITIES.BUSINESS,
         error: {
           error: error,
-          message: errorMessage,
+          message: errorMessage || error,
           title: error.name || error,
         },
       };
@@ -196,15 +197,15 @@ class WzRulesetEditor extends Component {
   };
 
   render() {
-    const { section, addingRulesetFile, fileContent } = this.props.state;
+    const { section, addingFile, fileContent } = this.props;
     const { wazuhNotReadyYet } = this.props;
     const { name, content, path, showWarningRestart } = this.state;
     const isRules = path.includes('rules') ? 'Ruleset Test' : 'Decoders Test';
 
-    const isEditable = addingRulesetFile
+    const isEditable = addingFile
       ? true
       : path !== 'ruleset/rules' && path !== 'ruleset/decoders';
-    let nameForSaving = addingRulesetFile ? this.state.inputValue : name;
+    let nameForSaving = addingFile ? this.state.inputValue : name;
     nameForSaving = nameForSaving.endsWith('.xml') ? nameForSaving : `${nameForSaving}.xml`;
     const overwrite = fileContent ? true : false;
 
@@ -212,7 +213,6 @@ class WzRulesetEditor extends Component {
 
     const onClickOpenLogtest = () => {
       this.props.logtestProps.openCloseFlyout();
-      this.props.showFlyoutLogtest(true);
     };
 
     const buildLogtestButton = () => {
@@ -262,7 +262,7 @@ class WzRulesetEditor extends Component {
             title="Unsubmitted changes"
             onConfirm={() => {
               closeModal;
-              this.props.cleanInfo();
+              this.props.cleanEditState();
             }}
             onCancel={closeModal}
             cancelButtonText="No, don't do it"
@@ -300,7 +300,7 @@ class WzRulesetEditor extends Component {
                                 ) {
                                   showModal();
                                 } else {
-                                  this.props.cleanInfo();
+                                  this.props.cleanEditState();
                                 }
                               }}
                             />
@@ -332,7 +332,7 @@ class WzRulesetEditor extends Component {
                                 ) {
                                   showModal();
                                 } else {
-                                  this.props.cleanInfo();
+                                  this.props.cleanEditState();
                                 }
                               }}
                             />
@@ -408,7 +408,6 @@ class WzRulesetEditor extends Component {
 
 const mapStateToProps = state => {
   return {
-    state: state.rulesetReducers,
     wazuhNotReadyYet: state.appStateReducers.wazuhNotReadyYet,
     showFlyout: state.appStateReducers.showFlyoutLogtest,
   };
@@ -416,14 +415,11 @@ const mapStateToProps = state => {
 
 const mapDispatchToProps = dispatch => {
   return {
-    cleanInfo: () => dispatch(cleanInfo()),
-    updateFileContent: content => dispatch(updateFileContent(content)),
     updateWazuhNotReadyYet: wazuhNotReadyYet => dispatch(updateWazuhNotReadyYet(wazuhNotReadyYet)),
-    showFlyoutLogtest: showFlyout => dispatch(showFlyoutLogtest(showFlyout)),
   };
 };
 
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(WzRulesetEditor);
+)(WzFileEditor);