+++ public/redux/reducers/rootReducers.js
@@ -11,7 +11,6 @@
  */
 
 import { combineReducers } from 'redux';
-import rulesetReducers from './rulesetReducers';
 import groupsReducers from './groupsReducers';
 import statusReducers from './statusReducers';
 import reportingReducers from './reportingReducers';
@@ -26,7 +25,6 @@ import appConfig from './appConfigReducers';
 import restartWazuhReducers from './restartReducers'
 
 export default combineReducers({
-  rulesetReducers,
   groupsReducers,
   statusReducers,
   reportingReducers,